Refactor HomeViewController to use ScoreView
Fixes #11
